desktop_app/requirements.txt
PySide6>=6.7
requests>=2.32
python-dotenv>=1.0
pillow>=10.4
pypdfium2>=4.26.0
pikepdf>=8.10.0